requirements.txt
beautifulsoup4==4.13.4
chromadb==0.6.3
fastapi==0.115.12
html2text==2024.2.26
langchain==0.3.23
langchain_community==0.3.21
langchain_core==0.3.52
langchain_openai==0.3.13
pandas==2.2.3
pydantic==2.11.3
python-dotenv==1.1.0
Requests==2.32.3
schedule==1.2.2
selenium==4.31.0
sentence_transformers==3.4.1
webdriver_manager==4.0.2
